requirements.txt
pyshark==0.6
pyinstaller==6.4.0
Pillow==10.2.0
tqdm==4.66.2
